Implemented state and decorator pattern
Wrote tests for it but code coverage went down, most likely because of
update to a rc of scala.
The update was required because the test tool had a runtime exception
on the older version.
Line 80 in ControllerSpec needs to be taken care of, it somehow opens
another field sometimes.