TOML format for FileIO implemented and improvements to other formats.
State attribute removed from the files. The state attribute can be gained from opening all openFields again.
In case the user changes the mines, the game needs to calculate the right points for the Position again.
The tests were rewritten to conform to this change.
FileIO also has a general `verifyData` function to check various constraints in the file. Error handling is more elaborate with this change.
Also finally fixed the GUI not resizing on new or loading game.